+++ shoppingassistant/src/main/java/org/shoppingassistant/advanced/ShoppingListItem.java
@@ -5,21 +5,21 @@ import org.shoppingassistant.ShoppingItem;
 
 public class ShoppingListItem extends ShoppingItem {
 
-    Integer shoppingListId;
+    Long shoppingListId;
 
     public ShoppingListItem() {
     }
 
-    public ShoppingListItem(Integer id, Integer shoppingListId, ItemCategory category, String description, int quantity) {
+    public ShoppingListItem(Long id, Long shoppingListId, ItemCategory category, String description, int quantity) {
         super(id, category, description, quantity);
         this.shoppingListId = shoppingListId;
     }
 
-    public Integer getShoppingListId() {
+    public Long getShoppingListId() {
         return shoppingListId;
     }
 
-    public void setShoppingListId(Integer shoppingListId) {
+    public void setShoppingListId(Long shoppingListId) {
         this.shoppingListId = shoppingListId;
     }
